support default integer size 32bit calculation

+++ src/e0after_tra.F90
@@ -10,6 +10,7 @@ SUBROUTINE e0aftertra
 ! +=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=
 ! +=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_file_manager
     use module_global_variables
     use module_index_utils, only: convert_global_to_active_idx
@@ -19,7 +20,7 @@ SUBROUTINE e0aftertra
 #endif
     Implicit NONE
 
-    integer :: ii, jj, kk, ll
+    integer(kind=int64) :: ii, jj, kk, ll
     integer :: j, i, k, l
 
     real(8) :: dr, di
